fix: 보관한, 알약 상세, 검색결과 화면 무한 리렌더링 버그 - 분리된 커스텀훅과 사용하는 컴포넌트의 useEffect의 문제로 인해 무한 리렌더링 발생 - 해당 커스텀훅 삭제, useFocusEffect 삭제, 분리된 useEffect 통합
